more stupid crap
git-svn-id: svn://svn.mplayerhq.hu/mplayer/trunk@12387 b3059339-0415-0410-9bf9-f77b7e298cf2

+++ libaf/af_format.c
@@ -234,7 +234,7 @@ static void uninit(struct af_instance_s* af)
 {
   if(af->data)
     free(af->data);
-  (int)af->setup = 0;  
+  af->setup = 0;  
 }
 
 // Filter data through filter
